Extract a base for node value views
